c backend float to double

+++ core/src/codegen/c/code_gen_c.rs
@@ -112,7 +112,7 @@ impl CodeGenC {
                 BuiltinTypedTypeKind::Integer => "int".to_string(),
                 BuiltinTypedTypeKind::Boolean => "char".to_string(),
                 BuiltinTypedTypeKind::Char => "char*".to_string(),
-                BuiltinTypedTypeKind::Float => "float".to_string(),
+                BuiltinTypedTypeKind::Float => "double".to_string(),
                 BuiltinTypedTypeKind::Lambda {
                     parameters,
                     return_type,
@@ -151,7 +151,7 @@ impl CodeGenC {
                 BuiltinTypedTypeKind::Integer => "int".to_string(),
                 BuiltinTypedTypeKind::Boolean => "char".to_string(),
                 BuiltinTypedTypeKind::Char => "char*".to_string(),
-                BuiltinTypedTypeKind::Float => "float".to_string(),
+                BuiltinTypedTypeKind::Float => "double".to_string(),
                 BuiltinTypedTypeKind::Lambda {
                     parameters: _,
                     return_type: _,
